change: automatically fill codec info during msg server registration
