+++ baseapp/msg_service_router.go
@@ -3,6 +3,7 @@ package baseapp
 import (
 	"context"
 	"fmt"
+	"strings"
 
 	gogogrpc "github.com/gogo/protobuf/grpc"
 	"github.com/gogo/protobuf/proto"
@@ -41,32 +42,27 @@ func (msr *MsgServiceRouter) Handler(methodName string) MsgServiceHandler {
 // service description, handler is an object which implements that gRPC service.
 //
 // This function PANICs:
-// - if it is called before the service `Msg`s have been registered using
-//   RegisterInterfaces,
+// - if interface registry is nil
 // - or if a service is being registered twice.
 func (msr *MsgServiceRouter) RegisterService(sd *grpc.ServiceDesc, handler interface{}) {
+	// register the request types in the interface registry
+	err := msr.registerInputs(sd)
+	if err != nil {
+		panic(err)
+	}
+	// map requests to handlers
+	err = msr.registerMethods(sd, handler)
+	if err != nil {
+		panic(err)
+	}
+}
+
+// registerMethods will register the handlers for each input in the given grpc.ServiceDesc
+func (msr *MsgServiceRouter) registerMethods(sd *grpc.ServiceDesc, handler interface{}) error {
 	// Adds a top-level query handler based on the gRPC service name.
 	for _, method := range sd.Methods {
 		fqMethod := fmt.Sprintf("/%s/%s", sd.ServiceName, method.MethodName)
 		methodHandler := method.Handler
-
-		// Check that the service Msg fully-qualified method name has already
-		// been registered (via RegisterInterfaces). If the user registers a
-		// service without registering according service Msg type, there might be
-		// some unexpected behavior down the road. Since we can't return an error
-		// (`Server.RegisterService` interface restriction) we panic (at startup).
-		serviceMsg, err := msr.interfaceRegistry.Resolve(fqMethod)
-		if err != nil || serviceMsg == nil {
-			panic(
-				fmt.Errorf(
-					"type_url %s has not been registered yet. "+
-						"Before calling RegisterService, you must register all interfaces by calling the `RegisterInterfaces` "+
-						"method on module.BasicManager. Each module should call `msgservice.RegisterMsgServiceDesc` inside its "+
-						"`RegisterInterfaces` method with the `_Msg_serviceDesc` generated by proto-gen",
-					fqMethod,
-				),
-			)
-		}
 
 		// Check that each service is only registered once. If a service is
 		// registered more than once, then we should error. Since we can't
@@ -74,12 +70,10 @@ func (msr *MsgServiceRouter) RegisterService(sd *grpc.ServiceDesc, handler inter
 		// panic (at startup).
 		_, found := msr.routes[fqMethod]
 		if found {
-			panic(
-				fmt.Errorf(
-					"msg service %s has already been registered. Please make sure to only register each service once. "+
-						"This usually means that there are conflicting modules registering the same msg service",
-					fqMethod,
-				),
+			return fmt.Errorf(
+				"msg service %s has already been registered. Please make sure to only register each service once. "+
+					"This usually means that there are conflicting modules registering the same msg service",
+				fqMethod,
 			)
 		}
 
@@ -104,11 +98,55 @@ func (msr *MsgServiceRouter) RegisterService(sd *grpc.ServiceDesc, handler inter
 			return sdk.WrapServiceResult(ctx, resMsg, err)
 		}
 	}
+	return nil
 }
 
-// SetInterfaceRegistry sets the interface registry for the router.
-func (msr *MsgServiceRouter) SetInterfaceRegistry(interfaceRegistry codectypes.InterfaceRegistry) {
-	msr.interfaceRegistry = interfaceRegistry
+// registerInputs registers the requests of a MsgServer service descriptor
+// in the interface registry to allow BaseApp to get the correct requests
+// from the Deliver/CheckTx implementation.
+func (msr *MsgServiceRouter) registerInputs(sd *grpc.ServiceDesc) error {
+	if msr.interfaceRegistry == nil {
+		return fmt.Errorf("nil interface registry")
+	}
+	// Adds a top-level type_url based on the Msg service name.
+	for _, method := range sd.Methods {
+		fqMethod := fmt.Sprintf("/%s/%s", sd.ServiceName, method.MethodName)
+		methodHandler := method.Handler
+
+		// NOTE: This is how we pull the concrete request type for each handler for registering in the InterfaceRegistry.
+		// This approach is maybe a bit hacky, but less hacky than reflecting on the handler object itself.
+		// We use a no-op interceptor to avoid actually calling into the handler itself.
+		_, _ = methodHandler(nil, context.Background(), func(i interface{}) error {
+			msg, ok := i.(proto.Message)
+			if !ok {
+				// We panic here because there is no other alternative and the app cannot be initialized correctly
+				// this should only happen if there is a problem with code generation in which case the app won't
+				// work correctly anyway.
+				panic(fmt.Errorf("can't register request type %T for service method %s", i, fqMethod))
+			}
+
+			msr.interfaceRegistry.RegisterCustomTypeURL((*sdk.Msg)(nil), fqMethod, msg)
+			return nil
+		}, noopInterceptor)
+
+	}
+
+	return nil
+}
+
+func (msr *MsgServiceRouter) SetInterfaceRegistry(ir codectypes.InterfaceRegistry) {
+	msr.interfaceRegistry = ir
+}
+
+// gRPC NOOP interceptor
+func noopInterceptor(_ context.Context, _ interface{}, _ *grpc.UnaryServerInfo, _ grpc.UnaryHandler) (interface{}, error) {
+	return nil, nil
+}
+
+// IsServiceMsg checks if a type URL corresponds to a service method name,
+// i.e. /cosmos.bank.Msg/Send vs /cosmos.bank.MsgSend
+func IsServiceMsg(typeURL string) bool {
+	return strings.Count(typeURL, "/") >= 2
 }
 
 func noopDecoder(_ interface{}) error { return nil }
