+++ x/authz/types/generic_authorization.go
@@ -2,8 +2,6 @@ package types
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-	"github.com/cosmos/cosmos-sdk/types/msgservice"
 	"github.com/cosmos/cosmos-sdk/x/authz/exported"
 )
 
@@ -30,8 +28,11 @@ func (authorization GenericAuthorization) Accept(ctx sdk.Context, msg sdk.Servic
 
 // ValidateBasic implements Authorization.ValidateBasic.
 func (authorization GenericAuthorization) ValidateBasic() error {
-	if !msgservice.IsServiceMsg(authorization.MessageName) {
-		return sdkerrors.Wrapf(sdkerrors.ErrInvalidType, " %s is not a valid service msg", authorization.MessageName)
-	}
+	// TODO: now that they're basically the same how do we handle this
+	/*
+		if !msgservice.IsServiceMsg(authorization.MessageName) {
+			return sdkerrors.Wrapf(sdkerrors.ErrInvalidType, " %s is not a valid service msg", authorization.MessageName)
+		}
+	*/
 	return nil
 }
