+++ x/evidence/types/codec.go
@@ -5,7 +5,6 @@ import (
 	"github.com/cosmos/cosmos-sdk/codec/types"
 	cryptocodec "github.com/cosmos/cosmos-sdk/crypto/codec"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/types/msgservice"
 	"github.com/cosmos/cosmos-sdk/x/evidence/exported"
 )
 
@@ -24,8 +23,6 @@ func RegisterInterfaces(registry types.InterfaceRegistry) {
 		(*exported.Evidence)(nil),
 		&Equivocation{},
 	)
-
-	msgservice.RegisterMsgServiceDesc(registry, &_Msg_serviceDesc)
 }
 
 var (
