+++ x/feegrant/types/codec.go
@@ -3,7 +3,6 @@ package types
 import (
 	"github.com/cosmos/cosmos-sdk/codec/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/types/msgservice"
 )
 
 // RegisterInterfaces registers the interfaces types with the interface registry
@@ -20,6 +19,4 @@ func RegisterInterfaces(registry types.InterfaceRegistry) {
 		&PeriodicFeeAllowance{},
 		&AllowedMsgFeeAllowance{},
 	)
-
-	msgservice.RegisterMsgServiceDesc(registry, &_Msg_serviceDesc)
 }
