+++ x/gov/types/codec.go
@@ -5,7 +5,6 @@ import (
 	"github.com/cosmos/cosmos-sdk/codec/types"
 	cryptocodec "github.com/cosmos/cosmos-sdk/crypto/codec"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/types/msgservice"
 )
 
 // RegisterLegacyAminoCodec registers all the necessary types and interfaces for the
@@ -31,8 +30,6 @@ func RegisterInterfaces(registry types.InterfaceRegistry) {
 		(*Content)(nil),
 		&TextProposal{},
 	)
-
-	msgservice.RegisterMsgServiceDesc(registry, &_Msg_serviceDesc)
 }
 
 // RegisterProposalTypeCodec registers an external proposal content type defined
